refactoring vol 2
