+++ src/types.ts
@@ -1,1 +1,3 @@
 export type Setter<T> = React.Dispatch<React.SetStateAction<T>>;
+
+export type Tile = { z: number; x: number; y: number };
